Add CMake install target along with build options
Following CMake options have been added with the defaults specified:

* -DWITH_PIC=OFF (Compile static library as PIC too)
* -DBUILD_STATIC_LIBS=ON
* -DBUILD_SHARED_LIBS=ON
* -DBUILD_MANUAL_TEST=ON
* -DBUILD_MACOS_FATLIB=ON
