+++ XInputSimulator/.idea/workspace.xml
@@ -87,16 +87,6 @@
           </provider>
         </entry>
       </file>
-      <file leaf-file-name="xinputsimulatorimpllinux.h" pinned="false" current-in-tab="false">
-        <entry file="file://$PROJECT_DIR$/xinputsimulatorimpllinux.h">
-          <provider selected="true" editor-type-id="text-editor">
-            <state vertical-scroll-proportion="0.0">
-              <caret line="0" column="0" selection-start-line="0" selection-start-column="0" selection-end-line="0" selection-end-column="0" />
-              <folding />
-            </state>
-          </provider>
-        </entry>
-      </file>
       <file leaf-file-name="notimplementedexception.cpp" pinned="false" current-in-tab="false">
         <entry file="file://$PROJECT_DIR$/notimplementedexception.cpp">
           <provider selected="true" editor-type-id="text-editor">
