+++ XInputSimulator/XInputSimulator.pc.in
@@ -1,0 +1,14 @@
+prefix=@CMAKE_INSTALL_PREFIX@
+exec_prefix=${prefix}
+libdir=${exec_prefix}/lib
+includedir=${prefix}/include
+
+Name: XInputSimulator
+Description: Cross-Platform Simulator mouse and keyboard input
+URL: https://github.com/pythoneer/XInputSimulator
+Version: @PROJECT_VERSION@
+Requires:
+Libs: -L${libdir} -lXInputSimulator
+Libs.private: @PKG_CONFIG_EXTRA_LIBS@
+Cflags: -I${includedir}
+
