+++ XInputSimulator/notimplementedexception.cpp
@@ -15,7 +15,7 @@
 //    You should have received a copy of the GNU Lesser Public License
 //    along with XInputSimulator.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "notimplementedexception.h"
+#include "xinputsimulator.h"
 
 NotImplementedException::NotImplementedException()
     :
