--- XInputSimulator/notimplementedexception.h
@@ -1,29 +1,0 @@
-//    Copyright 2013 Dustin Bensing
-
-//    This file is part of XInputSimulator.
-
-//    XInputSimulator is free software: you can redistribute it and/or modify
-//    it under the terms of the GNU Lesser Public License as published by
-//    the Free Software Foundation, either version 3 of the License, or
-//    any later version.
-
-//    XInputSimulator is distributed in the hope that it will be useful,
-//    but WITHOUT ANY WARRANTY; without even the implied warranty of
-//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//    GNU Lesser Public License for more details.
-
-//    You should have received a copy of the GNU Lesser Public License
-//    along with XInputSimulator.  If not, see <http://www.gnu.org/licenses/>.
-
-#ifndef NOTIMPLEMENTEDEXCEPTION_H
-#define NOTIMPLEMENTEDEXCEPTION_H
-
-#include <stdexcept>
-
-class NotImplementedException : public std::runtime_error
-{
-public:
-    NotImplementedException();
-};
-
-#endif // NOTIMPLEMENTEDEXCEPTION_H
