+++ XInputSimulator/xinputsimulatorimplmacos.cpp
@@ -23,8 +23,8 @@
 //sleep
 #include <unistd.h>
 
+#include "xinputsimulator.h"
 #include "xinputsimulatorimplmacos.h"
-#include "notimplementedexception.h"
 
 //#include <QDebug>
